refactor: move argument logic to different file
