+++ src/bcc2.c
@@ -10,6 +10,7 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 
+#include "args.h"
 #include "helper.h"
 #include "ir_gen.h"
 #include "lexer.h"
@@ -17,105 +18,10 @@
 #include "semantics.h"
 #include "ssa.h"
 
-struct Option {
-  const char *flag;
-  const char *description;
-  bool long_flag;
-
-  enum {
-    ARG_NONE,
-    ARG_REQUIRED,
-    ARG_OPTIONAL,
-  } required_arg;
-
-  enum {
-    OPT_BOOLEAN,
-    OPT_STRING,
-    OPT_INT,
-  } type;
-
-  union {
-    bool enabled;
-    const char *string;
-    int integer;
-  } out;
-};
-
-bool
-fill_flag_argument(const char *value, struct Option *opt) {
-  bool argument_used = false;
-  if (opt->required_arg == ARG_NONE && value != NULL) {
-    log_err_final("option '%s' doesn't allow an argument", opt->flag);
-  } else if (opt->required_arg == ARG_REQUIRED && value == NULL) {
-    log_err_final("option '%s' requires an argument", opt->flag);
-  }
-  switch (opt->type) {
-    case OPT_BOOLEAN:
-      opt->out.enabled = true;
-      break;
-    case OPT_STRING:
-      opt->out.string = value;
-      argument_used = true;
-      break;
-    case OPT_INT:
-      opt->out.integer = atoi(value);
-      argument_used = true;
-      break;
-  }
-  return argument_used;
-}
-
-void
-long_flag_parse(const char *flag, struct Option *opts[], size_t opts_size) {
-  for (size_t i = 0; i < opts_size; i++) {
-    if (opts[i]->long_flag &&
-        !strncmp(flag, opts[i]->flag, strlen(opts[i]->flag))) {
-      char *value;
-      if ((value = strchr(flag, '=')) != NULL) {
-        value++;
-      }
-      fill_flag_argument(value, opts[i]);
-      return;
-    }
-  }
-  log_err_final("unrecognized option %s", flag);
-}
-
-int
-short_flag_parse(const char *flag, char *next_argv, struct Option *opts[],
-                 size_t opts_size) {
-  for (size_t i = 0; i < opts_size; i++) {
-    if (!opts[i]->long_flag &&
-        !strncmp(flag, opts[i]->flag, strlen(opts[i]->flag))) {
-      char *value = next_argv;
-      return fill_flag_argument(value, opts[i]) + 1;
-    }
-  }
-  log_err_final("unrecognized option %s", flag);
-  return 0; // unreachable, just to stop a warning
-}
-
-void
-parse_args(int argc, char *argv[], struct Option *opts[], size_t opts_size,
-           char **input_file) {
-  for (int i = 1; i < argc;) {
-    if (argv[i][0] == '-') {
-      errno = 0;
-      if (argv[i][1] == '-') {
-        long_flag_parse(&argv[i][2], opts, opts_size);
-        i++;
-      } else {
-        i += short_flag_parse(&argv[i][1], argc - i != 1 ? argv[i + 1] : NULL,
-                              opts, opts_size);
-      }
-    } else {
-      if (*input_file != NULL) {
-        log_err_final("can't specify more than one input file");
-      }
-      *input_file = argv[i];
-      i++;
-    }
-  }
+void print_help(const char *program_name, const char *program_description, struct Option *opts[], size_t opts_size) {
+  printf("%s: %s\n", program_name, program_description);
+  printf("Usage: %s [OPTION]... [FILE]\n", program_name);
+  print_flags(opts, opts_size);
 }
 
 int
@@ -158,18 +64,18 @@ main(int argc, char *argv[]) {
 
   struct Option *opts[] = {&help, &version, &ast_dump_flag, &ir_dump_flag,
                            &reg_dump_flag};
+  size_t number_opts = sizeof(opts)/sizeof(struct Option *);
   char *in_filename = NULL;
 
-  parse_args(argc, argv, opts, sizeof(opts), &in_filename);
+  parse_args(argc, argv, opts, number_opts, &in_filename);
+  printf("%s", in_filename);
 
   if (!ir_dump_flag.out.enabled && reg_dump_flag.out.enabled) {
     log_err_final("cannot print registers without printing the IR");
   }
 
   if (help.out.enabled) {
-    printf("-h : prints help\n-v : prints version\n-ast : dumps ast to "
-           "stdout\n-ir : dumps ir to stdout\n-regs : dumps registers to "
-           "stdout\n");
+    print_help(argv[0], "Take two on compiler for beans.", opts, number_opts);
     exit(EXIT_SUCCESS);
   }
   if (version.out.enabled) {
